Run all spring-boot-2-starter endpoint tests in a single shared context
Previously each of the 13 *EndpointTest classes declared its own
@SpringBootTest with per-class imports, properties and nested
@TestConfiguration classes, so every class forked a separate Spring
application context.

All endpoint tests now extend AbstractEndpointIntegrationTest, which
carries the union of the previously per-class configuration. Since the
subclasses declare no context-affecting annotations of their own, they
produce an identical merged context configuration and share one cached
application context for the whole test run.

Details:
- Nested @TestConfiguration classes moved to top-level configs in their
  packages, since nested configs are auto-detected and change the
  context cache key.
- Conflicting bean definitions merged: a single SmartSanitizingFunction
  with the union sanitization list, a single GitInformationProvider /
  LibraryInformationProvider (details config, reused by the metadata
  test), and a single AxelixTestProperties replacing the three
  @ConfigurationProperties classes that all bound 'axelix.prop.test'.
- The three conflicting 'management.endpoints.web.exposure.include'
  values replaced with '*' in the shared configuration.
- AxelixLoggersEndpointTest establishes its logger-level baseline in
  @BeforeEach, since contexts of other tests reinitialize the global
  logging system after the shared context starts.
- TransactionMonitoringEndpointTest allows a 2 ms tolerance on the
  reconstructed end timestamps: with a warm shared context transactions
  complete in under a millisecond, exposing the independent truncation
  of the transaction and query durations.
