+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/beans/AxelixBeansEndpointTest.java
@@ -22,23 +22,12 @@ import java.util.function.Supplier;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.actuate.autoconfigure.condition.ConditionsReportEndpoint;
-import org.springframework.boot.actuate.beans.BeansEndpoint;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.TestPropertySource;
 
 import com.axelixlabs.axelix.common.api.BeansFeed;
 import com.axelixlabs.axelix.common.domain.http.HttpMethod;
-import com.axelixlabs.axelix.sbs.spring.core.auth.JwtAuthTestConfiguration;
-import com.axelixlabs.axelix.sbs.spring.core.conditions.ConditionalBeanRefBuilder;
-import com.axelixlabs.axelix.sbs.spring.core.conditions.DefaultConditionalBeanRefBuilder;
+import com.axelixlabs.axelix.sbs.spring.core.AbstractEndpointIntegrationTest;
+import com.axelixlabs.axelix.sbs.spring.core.utils.AxelixTestProperties;
 import com.axelixlabs.axelix.sbs.spring.core.utils.TestRestTemplateBuilder;
 import com.axelixlabs.axelix.sbs.spring.core.utils.auth.ProtectedEndpointTests;
 
@@ -50,68 +39,10 @@ import static org.assertj.core.api.InstanceOfAssertFactories.type;
  *
  * @author Mikhail Polivakha
  */
-@SpringBootTest(
-        classes = AxelixBeansEndpointTest.CurrentConfiguration.class,
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestPropertySource(properties = {"axelix.prop.test.name=axelix-beans"})
-@Import({BeansEndpoint.class, AxelixBeansEndpoint.class, ConditionsReportEndpoint.class, JwtAuthTestConfiguration.class
-})
-class AxelixBeansEndpointTest {
+class AxelixBeansEndpointTest extends AbstractEndpointIntegrationTest {
 
     @Autowired
     private TestRestTemplateBuilder testRestTemplate;
-
-    @TestConfiguration(value = "testCurrentConfiguration")
-    @EnableConfigurationProperties(AxelixPropTest.class)
-    static class CurrentConfiguration {
-
-        static final String QUALIFIERS_PERSISTENCE_POST_PROCESSOR = "qualifiersPersistencePostProcessor";
-        static final String BEAN_META_INFO_EXTRACTOR = "beanMetaInfoExtractor";
-        static final String CUSTOM_SUPPLIER = "customSupplier";
-
-        @Bean
-        public ConditionalBeanRefBuilder conditionalBeanRefBuilder() {
-            return new DefaultConditionalBeanRefBuilder();
-        }
-
-        @Bean
-        public BeansFeedBuilder testBeansFeedBuilder(
-                BeanMetaInfoExtractor beanMetaInfoExtractor,
-                ConfigurableApplicationContext configurableApplicationContext) {
-            return new DefaultBeansFeedBuilder(beanMetaInfoExtractor, configurableApplicationContext);
-        }
-
-        @Bean(BEAN_META_INFO_EXTRACTOR)
-        public BeanMetaInfoExtractor beanMetaInfoExtractor(
-                ConfigurableApplicationContext configurableApplicationContext,
-                ConditionalBeanRefBuilder conditionalBeanRefBuilder) {
-            return new DefaultBeanMetaInfoExtractor(configurableApplicationContext, conditionalBeanRefBuilder);
-        }
-
-        @Bean(QUALIFIERS_PERSISTENCE_POST_PROCESSOR)
-        public static QualifiersPersistencePostProcessor qualifiersPersistencePostProcessor() {
-            return new QualifiersPersistencePostProcessor();
-        }
-
-        @Bean(CUSTOM_SUPPLIER)
-        public Supplier<String> customSupplier() {
-            return () -> "value";
-        }
-    }
-
-    @ConfigurationProperties(prefix = "axelix.prop.test")
-    static class AxelixPropTest {
-
-        private String name;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-    }
 
     @Test
     void shouldReturnEnrichedBeansFeed() {
@@ -135,15 +66,17 @@ class AxelixBeansEndpointTest {
     void negativeAuthTests() {}
 
     private static void assertQualifiersPostProcessorBean(BeansFeed beanNameToBeanFeed) {
-        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, CurrentConfiguration.QUALIFIERS_PERSISTENCE_POST_PROCESSOR);
+        BeansFeed.Bean bean =
+                getBean(beanNameToBeanFeed, BeansEndpointTestConfiguration.QUALIFIERS_PERSISTENCE_POST_PROCESSOR);
 
         assertThat(bean.getBeanSource()).isInstanceOf(BeansFeed.BeanMethod.class);
         assertThat(bean.getBeanSource())
                 .asInstanceOf(type(BeansFeed.BeanMethod.class))
                 .satisfies(beanMethod -> {
                     assertThat(beanMethod.getMethodName())
-                            .isEqualTo(CurrentConfiguration.QUALIFIERS_PERSISTENCE_POST_PROCESSOR);
-                    assertThat(beanMethod.getEnclosingClassName()).isEqualTo("CurrentConfiguration");
+                            .isEqualTo(BeansEndpointTestConfiguration.QUALIFIERS_PERSISTENCE_POST_PROCESSOR);
+                    assertThat(beanMethod.getEnclosingClassName())
+                            .isEqualTo(BeansEndpointTestConfiguration.class.getSimpleName());
                 });
         assertThat(bean.isConfigPropsBean()).isFalse();
         assertThat(bean.getAutoConfigurationRef()).isNull();
@@ -157,14 +90,16 @@ class AxelixBeansEndpointTest {
     }
 
     private static void assertBeanMetaInfoExtractor(BeansFeed beanNameToBeanFeed) {
-        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, CurrentConfiguration.BEAN_META_INFO_EXTRACTOR);
+        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, BeansEndpointTestConfiguration.BEAN_META_INFO_EXTRACTOR);
 
         assertThat(bean.getBeanSource()).isInstanceOf(BeansFeed.BeanMethod.class);
         assertThat(bean.getBeanSource())
                 .asInstanceOf(type(BeansFeed.BeanMethod.class))
                 .satisfies(beanMethod -> {
-                    assertThat(beanMethod.getMethodName()).isEqualTo(CurrentConfiguration.BEAN_META_INFO_EXTRACTOR);
-                    assertThat(beanMethod.getEnclosingClassName()).isEqualTo("CurrentConfiguration");
+                    assertThat(beanMethod.getMethodName())
+                            .isEqualTo(BeansEndpointTestConfiguration.BEAN_META_INFO_EXTRACTOR);
+                    assertThat(beanMethod.getEnclosingClassName())
+                            .isEqualTo(BeansEndpointTestConfiguration.class.getSimpleName());
                 });
         assertThat(bean.isConfigPropsBean()).isFalse();
         assertThat(bean.getAutoConfigurationRef()).isNull();
@@ -181,14 +116,15 @@ class AxelixBeansEndpointTest {
     }
 
     private static void assertCustomBeanSupplier(BeansFeed beanNameToBeanFeed) {
-        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, CurrentConfiguration.CUSTOM_SUPPLIER);
+        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, BeansEndpointTestConfiguration.CUSTOM_SUPPLIER);
 
         assertThat(bean.getBeanSource()).isInstanceOf(BeansFeed.BeanMethod.class);
         assertThat(bean.getBeanSource())
                 .asInstanceOf(type(BeansFeed.BeanMethod.class))
                 .satisfies(beanMethod -> {
-                    assertThat(beanMethod.getMethodName()).isEqualTo(CurrentConfiguration.CUSTOM_SUPPLIER);
-                    assertThat(beanMethod.getEnclosingClassName()).isEqualTo("CurrentConfiguration");
+                    assertThat(beanMethod.getMethodName()).isEqualTo(BeansEndpointTestConfiguration.CUSTOM_SUPPLIER);
+                    assertThat(beanMethod.getEnclosingClassName())
+                            .isEqualTo(BeansEndpointTestConfiguration.class.getSimpleName());
                 });
         assertThat(bean.isConfigPropsBean()).isFalse();
         assertThat(bean.getAutoConfigurationRef()).isNull();
@@ -202,9 +138,9 @@ class AxelixBeansEndpointTest {
     }
 
     private static void assertConfigPropsBeanName(BeansFeed beanNameToBeanFeed) {
-        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, AxelixPropTest.class.getName());
+        BeansFeed.Bean bean = getBean(beanNameToBeanFeed, AxelixTestProperties.class.getName());
 
-        assertThat(bean.getClassName()).isEqualTo(AxelixPropTest.class.getName());
+        assertThat(bean.getClassName()).isEqualTo(AxelixTestProperties.class.getName());
         assertThat(bean.getBeanSource()).isNotNull();
         assertThat(bean.isConfigPropsBean()).isTrue();
         assertThat(bean.getAutoConfigurationRef()).isNull();
