+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/master/MetadataEndpointTestConfiguration.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (C) 2025-2026 Axelix Labs
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.axelixlabs.axelix.sbs.spring.core.master;
+
+import java.lang.management.ManagementFactory;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+
+import com.axelixlabs.axelix.common.api.registration.BasicDiscoveryMetadata;
+import com.axelixlabs.axelix.common.domain.version.AxelixVersionDiscoverer;
+
+/**
+ * Test configuration for {@link AxelixMetadataEndpointTest}, part of the shared endpoint test
+ * context. The {@link GitInformationProvider} and {@link LibraryInformationProvider} beans are
+ * provided by {@link com.axelixlabs.axelix.sbs.spring.core.details.DetailsEndpointTestConfiguration}.
+ *
+ * @author Mikhail Polivakha
+ */
+@TestConfiguration
+@Import({AxelixMetadataEndpoint.class, DefaultServiceMetadataAssembler.class, CommitIdPluginShortBuildInfoProvider.class
+})
+public class MetadataEndpointTestConfiguration {
+
+    @Bean
+    VMFeaturesProvider vmFeaturesProvider() {
+        return new OptionsParsingVMFeaturesProvider(
+                ManagementFactory.getRuntimeMXBean().getInputArguments());
+    }
+
+    @Bean
+    HealthDetectionFunction healthDetectionFunction() {
+        return () -> BasicDiscoveryMetadata.HealthStatus.UP;
+    }
+
+    @Bean
+    AxelixVersionDiscoverer axelixVersionDiscoverer() {
+        return () -> "1.1.3";
+    }
+}
