+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/transactions/TransactionMonitoringEndpointTest.java
@@ -37,23 +37,15 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionTemplate;
 
-import com.axelixlabs.axelix.sbs.spring.core.auth.JwtAuthTestConfiguration;
-import com.axelixlabs.axelix.sbs.spring.core.transactions.TransactionMonitoringEndpointTest.TransactionMonitoringEndpointTestConfiguration;
+import com.axelixlabs.axelix.sbs.spring.core.AbstractEndpointIntegrationTest;
 import com.axelixlabs.axelix.sbs.spring.core.utils.TestRestTemplateBuilder;
 import com.axelixlabs.axelix.sbs.spring.core.utils.auth.ProtectedEndpointTests;
 
@@ -69,10 +61,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  * @author Nikita Kirillov
  * @author Sergey Cherkasov
  */
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestPropertySource(properties = {"management.endpoints.web.exposure.include=axelix-transactions-monitoring"})
-@Import({TransactionMonitoringEndpointTestConfiguration.class, JwtAuthTestConfiguration.class})
-class TransactionMonitoringEndpointTest {
+class TransactionMonitoringEndpointTest extends AbstractEndpointIntegrationTest {
+
+    // The end timestamps are reconstructed as 'startTimestampMs + duration', where the duration
+    // is truncated to whole milliseconds independently for the transaction and for each query.
+    // For sub-millisecond transactions the query end may therefore exceed the execution end by
+    // up to 2 ms, hence the tolerance.
+    private static final long TIMESTAMP_TOLERANCE_MS = 2;
 
     @Autowired
     private TestRestTemplateBuilder restTemplate;
@@ -132,8 +127,10 @@ class TransactionMonitoringEndpointTest {
 
         assertThat(executionStartTimestampMs).isLessThanOrEqualTo(executionEndTimestampMs);
         assertThat(queryStartTimestampMs).isLessThanOrEqualTo(queryEndTimestampMs);
-        assertThat(queryStartTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
-        assertThat(queryEndTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
+        assertThat(queryStartTimestampMs)
+                .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
+        assertThat(queryEndTimestampMs)
+                .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
 
         assertThatJson(responseBody).node("entrypoints[0].executionStats").isObject();
         assertThatJson(responseBody)
@@ -196,8 +193,10 @@ class TransactionMonitoringEndpointTest {
             Long queryEndTimestampMs = query.get("endTimestampMs").longValue();
 
             assertThat(queryStartTimestampMs).isLessThanOrEqualTo(queryEndTimestampMs);
-            assertThat(queryStartTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
-            assertThat(queryEndTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
+            assertThat(queryStartTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
+            assertThat(queryEndTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
         }
     }
 
@@ -254,8 +253,10 @@ class TransactionMonitoringEndpointTest {
             Long queryEndTimestampMs = query.get("endTimestampMs").longValue();
 
             assertThat(queryStartTimestampMs).isLessThanOrEqualTo(queryEndTimestampMs);
-            assertThat(queryStartTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
-            assertThat(queryEndTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
+            assertThat(queryStartTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
+            assertThat(queryEndTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
         }
     }
 
@@ -310,8 +311,10 @@ class TransactionMonitoringEndpointTest {
             Long queryEndTimestampMs = query.get("endTimestampMs").longValue();
 
             assertThat(queryStartTimestampMs).isLessThanOrEqualTo(queryEndTimestampMs);
-            assertThat(queryStartTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
-            assertThat(queryEndTimestampMs).isBetween(executionStartTimestampMs, executionEndTimestampMs);
+            assertThat(queryStartTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
+            assertThat(queryEndTimestampMs)
+                    .isBetween(executionStartTimestampMs, executionEndTimestampMs + TIMESTAMP_TOLERANCE_MS);
         }
     }
 
@@ -504,52 +507,6 @@ class TransactionMonitoringEndpointTest {
         return response.getBody();
     }
 
-    @TestConfiguration
-    @EnableJpaRepositories(basePackageClasses = OwnerRepository.class, considerNestedRepositories = true)
-    @EntityScan(basePackageClasses = {Owner.class, Pet.class})
-    static class TransactionMonitoringEndpointTestConfiguration {
-
-        @Bean
-        public TransactionMonitoringEndpoint transactionMonitoringEndpoint(
-                TransactionMonitoringService transactionMonitoringService) {
-            return new TransactionMonitoringEndpoint(transactionMonitoringService);
-        }
-
-        @Bean
-        public TransactionMonitoringService transactionMonitoringService(
-                TransactionStatsCollector transactionStatsCollector) {
-            return new DefaultTransactionMonitoringService(transactionStatsCollector);
-        }
-
-        @Bean
-        public TransactionStatsCollector transactionStatsCollector() {
-            return new DefaultTransactionStatsCollector(30);
-        }
-
-        @Bean
-        public TransactionMonitoringBeanPostProcessor transactionMonitoringBeanPostProcessor(
-                TransactionStatsCollector transactionStatsCollector, QueriesRecorder queriesCollector) {
-            return new TransactionMonitoringBeanPostProcessor(transactionStatsCollector, queriesCollector);
-        }
-
-        @Bean
-        public QueriesRecorder queriesStatsCollector() {
-            return new DefaultQueriesRecorder();
-        }
-
-        @Bean
-        public ProxyingDataSourceBeanPostProcessor transactionMonitoringDataSourceBeanPostProcessor(
-                QueriesRecorder queriesCollector) {
-            return new ProxyingDataSourceBeanPostProcessor(queriesCollector);
-        }
-
-        @Bean
-        public PropagationTestHelper propagationTestHelper(
-                OwnerRepository ownerRepository, @Lazy PropagationTestHelper self) {
-            return new PropagationTestHelper(ownerRepository, self);
-        }
-    }
-
     @Entity
     @Table(name = "owner")
     static class Owner {
@@ -617,21 +574,9 @@ class TransactionMonitoringEndpointTest {
     static class PropagationTestHelper {
 
         private final OwnerRepository ownerRepository;
-        private final PropagationTestHelper self;
 
-        public PropagationTestHelper(OwnerRepository ownerRepository, @Lazy PropagationTestHelper self) {
+        public PropagationTestHelper(OwnerRepository ownerRepository) {
             this.ownerRepository = ownerRepository;
-            this.self = self;
-        }
-
-        // IMPORTANT: Calling via 'self' proxy is required to properly test the REQUIRED -> REQUIRES_NEW stack behavior.
-        @Transactional(propagation = Propagation.REQUIRED)
-        public void outerRequiredMethod(String outerName) {
-            ownerRepository.save(new Owner().setLastName(outerName));
-
-            self.saveRequiresNew("SomeName");
-
-            ownerRepository.save(new Owner().setLastName(outerName));
         }
 
         @Transactional(propagation = Propagation.REQUIRES_NEW)
